MAGECLOUD-278 Put SCD in build step -more changes from previously working scd in build branch

+++ src/Magento/MagentoCloud/Console/Command/Build.php
@@ -65,32 +65,50 @@ class Build extends Command
 
     private function build()
     {
+        $this->env->setStaticDeployInBuild(false);
         $this->env->log("Start build.");
         $this->applyMccPatches();
         $this->applyCommittedPatches();
+        $this->composerDumpAutoload();
         $this->compileDI();
         $this->deployStaticContent();
         $this->clearInitDir();
         $this->env->execute('rm -rf app/etc/env.php');
         $this->env->execute('rm -rf app/etc/config.php');
+        $this->backupToInit();
+    }
 
-        /**
-         * Writable directories will be erased when the writable filesystem is mounted to them. This
-         * step backs them up to ./init/
-         */
+    /**
+     * Writable directories will be erased when the writable filesystem is mounted to them. This
+     * step backs them up to ./init/
+     */
+    private function backupToInit()
+    {
         $this->env->log("Copying writable directories to temp directory.");
         foreach ($this->env->writableDirs as $dir) {
             $this->env->execute(sprintf('mkdir -p init/%s', $dir));
             $this->env->execute(sprintf('mkdir -p %s', $dir));
-            $this->env->execute(sprintf('/bin/bash -c "shopt -s dotglob; cp -R %s/* ./init/%s/"', $dir, $dir));
-            $this->env->execute(sprintf('rm -rf %s', $dir));
-            $this->env->execute(sprintf('mkdir %s', $dir));
+
+            if (count(scandir($dir)) >  2) {
+                $this->env->execute(sprintf('/bin/bash -c "shopt -s dotglob; cp -R %s/* ./init/%s/"', $dir, $dir));
+                $this->env->execute(sprintf('rm -rf %s', $dir));
+                $this->env->execute(sprintf('mkdir -p %s', $dir));
+            }
         }
 
-        $this->env->execute('mkdir -p ./init/pub/static');
-        rename('./pub/static/', './init/pub/static');
-        mkdir('./pub/static/');
-
+        if ($this->env->isStaticDeployInBuild()) {
+            $this->env->log("Moving static content to init directory");
+            $this->env->execute('mkdir -p ./init/pub/');
+            if (file_exists('./init/pub/static')) {
+                $this->env->log("Remove ./init/pub/static");
+                unlink('./init/pub/static');
+            }
+            $this->env->execute('cp -R ./pub/static/ ./init/pub/static');
+            copy(
+                Environment::MAGENTO_ROOT . Environment::STATIC_CONTENT_DEPLOY_FLAG,
+                Environment::MAGENTO_ROOT . 'init/' . Environment::STATIC_CONTENT_DEPLOY_FLAG
+            );
+        }
     }
 
     public function deployStaticContent()
@@ -124,6 +142,7 @@ class Build extends Command
                 $this->env->execute(
                     "/usr/bin/php ./bin/magento setup:static-content:deploy $jobsOption $excludeThemesOptions $SCDLocales {$this->verbosityLevel}"
                 );
+                $this->env->setStaticDeployInBuild(true);
             } catch (\Exception $e) {
                 $this->env->log($e->getMessage());
             }
@@ -196,5 +215,10 @@ class Build extends Command
 
     private function getBuildOption($key) {
         return isset($this->buildOptions[$key]) ? $this->buildOptions[$key] : false;
+    }
+
+    private function composerDumpAutoload()
+    {
+        $this->env->execute('composer dump-autoload');
     }
 }
